feat: Add SystemInfoHandler to collect CPU name, frequency, and power, and integrate these metrics into cooler data reporting.

+++ src/WaterCoolerCLI/Models/CoolerData.cs
@@ -6,11 +6,11 @@
 
         public byte CpuVendor { get; set; }
         public byte CpuTemperature { get; set; }
-        public byte CpuFrequency { get; set; }
+        public int CpuFrequency { get; set; }
         public byte CpuThreadCount { get; set; }
         public byte CpuCoreCount { get; set; }
         public byte CpuUsage { get; set; }
-        public byte CpuPower { get; set; }
+        public int CpuPower { get; set; }
         public byte VRAMTemperature { get; set; }
         public byte LiqiudTemperature => 0;
     }
